* A standard clock device has been added.
* The "vmipstool" front end to compilation tools was added, and most of the random scripts lying around have been consolidated into it.

* The memory-mapped test device known as "TestDev" has been removed.

* Debian and Red Hat packaging files have been added.

* The installation procedure has been greatly improved.

* The VMIPS manual has been completed.

* Many, many bugs fixed as a result of a concerted attempt at testing everything.  Doubtless, some bugs still remain.

* Some portability fixes.

* Dumping of the stack (when registers are dumped) now works.

* Many strange messages the debugger backend used to print out have been quelled. Also, killing the debugger connection should no longer cause a crash.

* vmips now supports --help, --version and --print-config command-line options.

* The SPIM-compatible console device now works better when the host machine is byte-swapped with respect to the MIPS target.

+++ devicemap.cc
@@ -10,8 +10,8 @@ char *
 DeviceMap::descriptor_str(void)
 {
 	static char buff[80];
-	sprintf(buff,"(%lx)[b=%lx e=%lx t=dev n=%lx]\n->",this,getBase(),
-		getExtent(),getNext());
+	sprintf(buff,"(%lx)[b=%lx e=%lx t=dev n=%lx]\n->",(unsigned long)this,
+		getBase(), getExtent(),(unsigned long)getNext());
 	return buff;
 }
